Feat: Add Server Details Screen and navigation
This commit introduces the `ServerDetailsScreen` and integrates it into the navigation flow.

Key changes:
- Created `ServerDetailsScreen.kt` to display detailed information about a server.
- Added `ServerDetailsViewModel` to manage the state and logic for the details screen.
- Implemented `GetServerDetailsUseCase` to fetch server details from the `ServerDataSource`.
- Added a `getServerById` query to `RusthubDB.sq` and its corresponding implementation in `ServerDataSourceImpl`.
- Modified `ServerAction.OnServerClick` to use `Long` for server ID.
- Updated `KoinInitializer.kt` to provide `ServerDetailsViewModel` and `GetServerDetailsUseCase`.
- Adjusted `NavigationRoot.kt` to navigate to `ServerDetailsScreen` with the correct parameters.
- Updated `ServerScreen.kt` to pass `Long` for server ID when navigating.
- Introduced `ServerDetailsAction.kt` and `ServerDetailsState.kt` for the new screen.
- Updated `Destination.kt` to use `Long` for `id` in `ServerDetails`.

+++ shared/src/commonMain/kotlin/pl/cuyer/rusthub/presentation/features/ServerViewModel.kt
@@ -196,7 +196,7 @@ class ServerViewModel(
         clearFiltersUseCase()
     }
 
-    private fun navigateToServer(id: Int, name: String) {
+    private fun navigateToServer(id: Long, name: String) {
         coroutineScope.launch {
             _uiEvent.send(UiEvent.Navigate(ServerDetails(id, name)))
         }
